+++ sharding-orchestration/sharding-orchestration-core/src/main/java/io/shardingsphere/jdbc/orchestration/internal/OrchestrationFacade.java
@@ -21,18 +21,19 @@ import com.google.common.base.Preconditions;
 import io.shardingsphere.core.api.config.MasterSlaveRuleConfiguration;
 import io.shardingsphere.core.api.config.ShardingRuleConfiguration;
 import io.shardingsphere.core.jdbc.core.datasource.MasterSlaveDataSource;
-import io.shardingsphere.jdbc.orchestration.api.config.OrchestrationConfiguration;
+import io.shardingsphere.jdbc.orchestration.config.OrchestrationConfiguration;
+import io.shardingsphere.jdbc.orchestration.config.OrchestrationProxyConfiguration;
 import io.shardingsphere.jdbc.orchestration.internal.config.ConfigurationService;
 import io.shardingsphere.jdbc.orchestration.internal.listener.ListenerFactory;
 import io.shardingsphere.jdbc.orchestration.internal.state.datasource.DataSourceService;
 import io.shardingsphere.jdbc.orchestration.internal.state.instance.InstanceStateService;
-import io.shardingsphere.jdbc.orchestration.reg.api.RegistryCenter;
-import io.shardingsphere.jdbc.orchestration.reg.api.RegistryCenterConfiguration;
-import io.shardingsphere.jdbc.orchestration.reg.etcd.EtcdConfiguration;
-import io.shardingsphere.jdbc.orchestration.reg.etcd.EtcdRegistryCenter;
-import io.shardingsphere.jdbc.orchestration.reg.newzk.NewZookeeperRegistryCenter;
-import io.shardingsphere.jdbc.orchestration.reg.zookeeper.ZookeeperConfiguration;
-import io.shardingsphere.jdbc.orchestration.reg.zookeeper.ZookeeperRegistryCenter;
+import io.shardingsphere.orchestration.reg.api.RegistryCenter;
+import io.shardingsphere.orchestration.reg.api.RegistryCenterConfiguration;
+import io.shardingsphere.orchestration.reg.etcd.EtcdConfiguration;
+import io.shardingsphere.orchestration.reg.etcd.EtcdRegistryCenter;
+import io.shardingsphere.orchestration.reg.newzk.NewZookeeperRegistryCenter;
+import io.shardingsphere.orchestration.reg.zookeeper.ZookeeperConfiguration;
+import io.shardingsphere.orchestration.reg.zookeeper.ZookeeperRegistryCenter;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
